feat!: move branding to a top-level command with a positional form
`workos authkit branding {get,set}` becomes `workos branding {get,set}`.
Branding is not AuthKit-specific — the same record drives hosted AuthKit
pages and transactional emails — so it reads better as its own noun. The
old path was never released, so nothing is aliased.

`set` now also takes a slot and a file positionally, matching how the rest
of the CLI reads (`role get <slug>`, `organization get <orgId>`):

    workos branding set icon ./icon.png
    workos branding set logo-dark ./logo-dark.png

Slot names are the flag names, so the two forms agree. The flag form stays
for setting several images at once:

    workos branding set --logo ./logo.png --favicon ./favicon.ico

Mixing the forms in one invocation is rejected rather than merged: both
could name the same slot with different files, and silently picking a
winner would upload the wrong image.

Handlers move from `commands/authkit.ts` to `commands/branding.ts`, and the
smoke tier gains live coverage of the positional path (set one image, then
assert only that image moved).
