+++ src/utils/help-json.ts
@@ -395,79 +395,92 @@ const commands: CommandSchema[] = [
           },
         ],
       },
+    ],
+  },
+  {
+    name: 'branding',
+    description: 'Manage branding images (logo, icon, favicon) for an environment',
+    options: [insecureStorageOpt],
+    commands: [
       {
-        name: 'branding',
-        description: 'Manage AuthKit branding',
-        commands: [
+        name: 'get',
+        description: 'Show branding (logos, theme) for an environment',
+        options: [
           {
-            name: 'get',
-            description: 'Show AuthKit branding (logos, theme) for an environment',
-            options: [
-              {
-                name: 'environment-id',
-                type: 'string',
-                description: 'Environment ID (defaults to the active environment)',
-                required: false,
-                hidden: false,
-              },
-            ],
+            name: 'environment-id',
+            type: 'string',
+            description: 'Environment ID (defaults to the active environment)',
+            required: false,
+            hidden: false,
+          },
+        ],
+      },
+      {
+        name: 'set',
+        description: 'Upload branding images (logo, icon, favicon) for an environment',
+        positionals: [
+          {
+            name: 'slot',
+            type: 'string',
+            description: 'Which image to set (logo, logo-dark, icon, icon-dark, favicon, favicon-dark)',
+            required: false,
+          },
+          { name: 'file', type: 'string', description: 'Path to the image file', required: false },
+        ],
+        options: [
+          {
+            name: 'environment-id',
+            type: 'string',
+            description: 'Environment ID (defaults to the active environment)',
+            required: false,
+            hidden: false,
           },
           {
-            name: 'set',
-            description: 'Upload AuthKit branding images (logo, icon, favicon) for an environment',
-            options: [
-              {
-                name: 'environment-id',
-                type: 'string',
-                description: 'Environment ID (defaults to the active environment)',
-                required: false,
-                hidden: false,
-              },
-              {
-                name: 'logo',
-                type: 'string',
-                description: 'Path to the light-mode logo image',
-                required: false,
-                hidden: false,
-              },
-              {
-                name: 'logo-dark',
-                type: 'string',
-                description: 'Path to the dark-mode logo image',
-                required: false,
-                hidden: false,
-              },
-              {
-                name: 'icon',
-                type: 'string',
-                description: 'Path to the light-mode app icon image',
-                required: false,
-                hidden: false,
-              },
-              {
-                name: 'icon-dark',
-                type: 'string',
-                description: 'Path to the dark-mode app icon image',
-                required: false,
-                hidden: false,
-              },
-              {
-                name: 'favicon',
-                type: 'string',
-                description: 'Path to the light-mode favicon',
-                required: false,
-                hidden: false,
-              },
-              {
-                name: 'favicon-dark',
-                type: 'string',
-                description: 'Path to the dark-mode favicon',
-                required: false,
-                hidden: false,
-              },
-            ],
-            examples: ['workos authkit branding set --logo ./logo.png --icon ./icon.png --favicon ./favicon.ico'],
+            name: 'logo',
+            type: 'string',
+            description: 'Path to the light-mode logo image',
+            required: false,
+            hidden: false,
           },
+          {
+            name: 'logo-dark',
+            type: 'string',
+            description: 'Path to the dark-mode logo image',
+            required: false,
+            hidden: false,
+          },
+          {
+            name: 'icon',
+            type: 'string',
+            description: 'Path to the light-mode app icon image',
+            required: false,
+            hidden: false,
+          },
+          {
+            name: 'icon-dark',
+            type: 'string',
+            description: 'Path to the dark-mode app icon image',
+            required: false,
+            hidden: false,
+          },
+          {
+            name: 'favicon',
+            type: 'string',
+            description: 'Path to the light-mode favicon',
+            required: false,
+            hidden: false,
+          },
+          {
+            name: 'favicon-dark',
+            type: 'string',
+            description: 'Path to the dark-mode favicon',
+            required: false,
+            hidden: false,
+          },
+        ],
+        examples: [
+          'workos branding set icon ./icon.png',
+          'workos branding set --logo ./logo.png --icon ./icon.png --favicon ./favicon.ico',
         ],
       },
     ],
